1st change: New API structure. 2nd change: in python_app-> moved functions to their own file and call function from app.py depending on @app.route
