+++ node_app/views/flip.ejs
@@ -20,7 +20,7 @@
         </div>
         <div>
             <br><br><br>
-            <% if(articles != 0) { %>
+            <% if(response) { %>
                 <div>
                     <h1>Here is your flipped text:</h1>
                     <p><%= response %></p>
