+++ python_app/my-project-env/bin/pip3.8
@@ -1,0 +1,10 @@
+#!/bin/sh
+'''exec' "/mnt/c/Users/Stephanie M/Desktop/ScriptFlip/flip-the-script/python_app/my-project-env/bin/python3" "$0" "$@"
+' '''
+# -*- coding: utf-8 -*-
+import re
+import sys
+from pip._internal.cli.main import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
