+++ python_app/my-project-env/bin/python
@@ -1,0 +1,1 @@
+python3
