+++ python_app/my-project-env/bin/python3
@@ -1,0 +1,1 @@
+/usr/bin/python3
