+++ python_app/my-project-env/lib64
@@ -1,0 +1,1 @@
+lib
