+++ python_app/neutral.py
@@ -1,0 +1,212 @@
+import re
+from collections import Counter
+
+
+dict_neutral = {'United Kingdom': 'United Kingdom', 'Staunton': 'Staunton', 'taunt' : 'taunt', 'The Sportsman': 'The Sportsman,', #first exceptions
+                'THE SPORTSMAN': 'THE SPORTSMAN', 'Queens, New York' : 'Queens, New York', 'emotional intelligence': 'emotional intelligence', #first exceptions
+                'Emotional inteligence': 'Emotional inteligence', #first exceptions 
+                ' man\n' :  ' person', ' man ' :  ' person ', 'Man ': 'Person ', ' man.': ' person.', ' man,': ' person,', " man’": " person’", " man'": " person'", '-man' : '-person ', 
+                '-Man': '-Person', 'Man,': 'Person,', 'than a man': 'than a person',
+                ' men ': ' people ', 'Men ': 'People ', ' men.': ' people.', ' men,': ' people,', " men’": " people’", " men'": " people'", '-men': '-people', 
+                'Men,': 'People,', '-Men': '-People', 'than men': 'than people', "Men'": "People'", "Man'": "Person'",
+                ' woman\n': ' person', ' woman ': ' person ', 'Woman': 'Person', ' woman.': ' person.', ' woman,': ' person,', " woman’": " person’", " woman'": " person'", 
+                '-woman': '-person', '-Woman': '-Person', ' woman”': ' person”', ' Woman”': ' Person”', ' woman"': ' person"', ' Woman"': ' Person"', 'than a woman': 'than a person',
+                ' women ': ' people ', 'Women': 'People', ' women.': ' people.', ' women,': ' people,', " women’": " people’", " women'": " people'", '-women': '-people', 
+                '-Women': '-People', ' women”': ' people”', ' Women”': ' People”', ' women"': ' people"', ' Women"': ' People"', 'than women': 'than people', 
+                'Mr': 'Mx', 'Ms': 'Mx', 'Sir': 'Mx', 'Lady': 'Mx', 'Lord': 'Mx', ' lady': ' person', 'gentleman': 'person', 'ladies': 'people', 'gentlemen': 'people',  
+                'bride': 'nearlywed', 'groom ': 'nearlywed ', 'Bride': 'Nearlywed', 'Groom': 'Nearlywed', 
+                'fiancée': 'nearlywed', 'fiance': 'nearlywed', 'fiancee': 'nearlywed', 'Fiancée': 'Nearlywed', 'Fiance': 'Nearlywed', 'Fiancee': 'Nearlywed',
+                ' actor': ' performer', ' actors': ' performers', 'Actor': 'Performer', 'Actors': 'Performers', 'an actor': 'a performer', '-actor': '-performer',
+                'actress': 'performer', 'actresses': 'performers', 'Actress': 'Performer', 'Actresses': 'Performers', 'an actress': 'a performer', 
+                'assemblyman': 'member of assembly', 'assemblyman': 'Member of assembly', 'assemblymen': 'members of assembly', 'Assemblymen': 'Members of assembly',
+                'assemblywoman': 'member of asssembly', 'Assemblywoman': 'Member of assembly', 'assemblywomen': 'members of assembly', 'Assemblywomen': 'Members of assembly',
+                'businessman': 'business executive', 'Businessman': 'Business executive', 
+                'businessmen': 'business executives', 'Businessmen': 'Business executives', 
+                'businesswoman': 'business executive', 'Businesswoman': 'Business executive', 
+                'businesswomen': 'business executives', 'Businesswomen': 'Business executives', 
+                'heroine': 'hero', 'Heroine': 'Hero',
+                'cameraman': 'camera operator', 'Cameraman': 'Camera operator', 
+                'cameramen': 'camera operators', 'Cameramen': 'Camera operators', 
+                'camerawoman': 'camera operator', 'Camerawoman': 'Camera operator', 
+                'camerawomen': 'camera operators', 'Camerawomen': 'Camera operators', 
+                'chairman': 'chairperson', 'Chairman': 'Chairperson', 
+                'chairmen': 'chairpeople', 'Chairmen': 'Chairpeople', 
+                'chairwoman': 'chairperson', 'Chairwoman': 'Chairperson', 
+                'chairwomen': 'chairpeople', 'Chairwomen': 'Chairpeople', 
+                'congressman': 'member of congress', 'Congressman': 'Member of congress', 'congressmen': 'members of congress', 'Congressmen': 'Members of congress',
+                'congresswoman': 'member of congress', 'Congresswoman': 'Member of congress', 'congresswomen': 'members of congress', 'Congresswomen': 'Members of congress',
+                'councilman': 'council person', 'Councilman': 'Council person', 
+                'councilwoman': 'council person', 'Councilwoman': 'Council person', 
+                'fireman': 'firefighter', 'Fireman': 'Firefighter', 
+                'firemen': 'firefighters', 'Firemen': 'Firefighters', 
+                'freshman': 'first-year student', 'Freshman': 'First-year student', 
+                'freshmen ': 'first-year students', 'Freshmen ': 'First-year students', 
+                'milkmaid' : 'milker', 'Milkmaid': 'Milker',
+                'policeman': 'police officer', 'Policeman': 'Police officer', 
+                'policemen': 'police officers', 'Policemen': 'Police officers', 
+                'policewoman': 'police officer', 'Policewoman': 'Police officer', 
+                'policewomen': 'police officers', 'Policewomen': 'Police officers', 
+                'postman': 'postal worker', 'Postman': 'Postal worker',
+                'postmen': 'postal workers', 'Postmen': 'Postal workers',
+                'postwoman': 'postal worker', 'Postwoman': 'Postal worker',
+                'postwomen': 'postal workers', 'Postwomen': 'Postal workers',
+                'repairman' : 'repairer', 'Repairman': 'Repairer', 
+                'repairmen' : 'repairers', 'Repairmen': 'Repairers', 
+                'salesman': 'salesperson', 'Salesman': 'Salesperson',
+                'saleswoman': 'salesperson', 'Saleswoman': 'Salesperson',
+                'salesmen': 'salespeople', 'Salesmen': 'Salespeople',
+                'saleswomen': 'salespeople', 'Saleswomen': 'Salespeople',
+                'spokesman': 'spokesperson', 'Spokesman': 'Spokesperson',
+                'spokeswoman': 'spokesperson', 'Spokeswoman': 'Spokesperson',
+                'spokeswomen': 'spokespeople', 'Spokeswomen': 'Spokespeople',
+                'spokesmen': 'spokespeople', 'Spokesmen': 'Spokespeople',
+                'sportsman': 'athlete', 'Sportsman': 'Athlete',
+                'sportswoman': 'athlete', 'Sportswoman': 'Athlete',
+                'sportswomen': 'athletes', 'Sportswomen': 'Athletes',
+                'sportsmen': 'athletes', 'Sportsmen': 'Athletes',
+                'SPORTSMAN': 'ATHLETE',
+                'workman': 'worker', 'Workman': 'Worker',
+                'workmen': 'workers', 'Workmen': 'Workers',
+                ' he does': ' they do', 'He does': 'They do', 'Does he': 'Do they',
+                ' she does': ' they do', 'She does': 'They do', 'Does she': 'Do they',
+                ' he is': ' they are', ' he was': ' they were', 'Is he': 'Are they', 'Is she': 'Are they', ' he is.': ' they are.', ' he is?': ' they are?', 
+                ' he has': ' they have',
+                'He is ': 'They are ', 'He was ': 'They were ', '(he is': '(they are', '(he was': '(they were',
+                'He has ': 'They have ',
+                ' she is ': ' they are ', ' she was ': ' they were ', '(she is': '(they are', '(she was': '(they were', ' she is.': ' they are.', ' she is?': ' they are?',
+                ' she has ': ' they have ',
+                'She is ': 'They are ', 'She was ': 'They were ',
+                'She has ': 'They have ',
+                'She/he': 'They', 'she/he': 'they', 'he/she': 'they', 'He/she': 'they', 'He/She': 'They', 
+                'HerName': 'TheirName', 'HisName': 'TheirName',
+                "he’ll": "they’ll", "she’ll": "they’ll", "he'll": "they'll", "she'll": "they'll", 
+                'she’s read': 'they’ve read', "she's read": "they've read", 'She’s read': 'They’ve read', "She's read": "They've read",
+                'he’s read': 'they’ve read', "he's read": "they've read", 'He’s read': 'They’ve read', "He's read": "They've read",
+                "he’s": "they’re", "she’s": "they’re", "he’d": "they’d", "she’d": "they’d", "he's": "they're", "she's": "they're", "he'd": "they'd", "she'd": "they'd",             
+                "He’s": "They’re", "She’s": "They’re", "He’d": "They’d", "She’d": "They’d", "He's": "They're", "She's": "They're", "He'd": "They'd", "She'd": "They'd",
+                " he isn’t": " they aren’t", " she isn’t": " they aren’t", " he isn't": " they aren't", " she isn't": " they aren't",
+                "He isn’t": "They aren’t", "She isn’t": "They aren’t", "He isn't": "They aren't", "She isn't": "They aren't",
+                ' he still has ': ' they still have ', 'He still has ': 'They still have ', 
+                ' she still has ': ' they still have ', 'She still has ': 'They still have ', 
+                ' he ': ' they ', 'He ': 'They ', "He’ll": "They’ll", '(he ': '(they ',
+                ' she ': ' they ', 'She ': 'They ', "She’ll": "They’ll", '(she ': '(they ',
+                ' her over': ' them over', ' her by': ' them by', 
+                ' his ': ' their ', 'His are': 'Theirs are', 'His is': 'Theirs is', 'His was': 'Theirs was','His ': 'Their ', ' his.': ' theirs.', ' him,': ' them,',
+                ' her ': ' their ', 'Hers are': 'Theirs are', 'Hers were': 'Theirs were', 'Her is': 'Theirs is', 'Her was': 'Theirs was', 'Her ': 'Their ', ' hers.': ' theirs.', 
+                ' her out': ' them out', ' her,': ' them,', ' her “that' : ' them “that', ' her that' : ' them that', ' her "that' : ' them "that',
+                ' him ': ' them ', 'himself': 'themself', 'herself': 'themself', ' him.': ' them.', ' her.': ' them.', ' him boy': ' them kiddo', ' her girl': ' them kiddo', 
+                ' her up': ' them up', 'takes her home': 'takes them home', 'take her home': 'take them home', 
+                ' hers ': ' them ', 
+                'from her is': 'from them is', 'From her is': 'From them is',
+                'from her was': 'from them was', 'From her was': 'From them was',
+                ' her back to' : ' them back to', 
+                'female ': None, 'Female ': None, 'female-': None, 'Female-': None,
+                'feminine ': None, 'Feminine ': None, "‘feminine’ ": None, "‘Feminine’ ": None, "'feminine' ": None, "'Feminine' ": None, 'feminine, ': None,
+                'feminist ': 'gender-equality activist', 'Feminist ': 'Gender-equality activist',
+                'misandrist': 'gender-equality activist', 'Misandrist': 'Gender-equality activist',
+                'masculine ': None, 'Masculine ': None, "‘masculine’ ": None, "‘Masculine’ ": None, "'masculine' ": None, "'Masculine' ": None, 'masculine, ': None, 
+                'male ': None, 'Male ': None, 'male-': None, 'Male-': None,
+                'dudely ': None, 'Dudely ': None,
+                'womanly ': None, 'Womanly ': None,
+                'boyish ': None, 'Boyish ': None,
+                'girly ': None, 'Girly ': None,
+                'manly ': None, 'Manly ': None,
+                'the Rev. ': None,
+                'feminism': 'gender-equality activism', 'Feminism': 'Gender-equality activism', 
+                'misandry': 'gender-equality activism', 'Misandry': 'Gender-equality activism',
+                'femininity': 'humanity', 'masculinity': 'humanity', 'Femininity': 'Humanity', 'Masculinity': 'Humanity',
+                'bossy': 'assertive', 'pushy': 'assertive', 'Bossy': 'Assertive', 'Pushy': 'Assertive',
+                'emotional': 'empathetic', 'hormonal': 'passionate', 'Emotional': 'Empathetic', 'Hormonal': 'Passionate',
+                'ditsy': 'silly', 'frigid': 'lacking sexual responsivness', 'Ditsy': 'Silly', 'Frigid': 'Lacking sexual responsivness',
+                'frumpy': 'dowdy', 'Frumpy': 'Dowdy',
+                'shrill': 'high pitched', 'Shrill': 'High pitched',
+                'hysterical': 'irrational', 'Hysterical': 'Irrational',
+                'mumsy': 'old fashioned', 'Mumsy': 'Old fashioned',
+                'virile': 'energetic', 'Virile': 'Energetic',
+                'emperor': 'ruler', 'empress': 'ruler', 'Emperor': 'Ruler', 'Empress': 'Ruler',  
+                'queen': 'monarch', ' king ': ' monarch ', 'Queen': 'Monarch', 'King ': 'Monarch ', 'Kingdom': 'Realm', 'kingdom': 'realm', 'King’': 'Monarch’',
+                'king’': 'monarch’',
+                'princess': "monarch's child", 'Princess': "Monarch's child", 
+                'princesses': "monarch's children", 'Princesses': "Monarch's children", 
+                'prince ': "monarch's child ", 'Prince ': "Monarch's child ", 
+                'princes ': "monarch's children ", 'Princes ': "Monarch's children ", 
+                'manliness ': 'humanness',
+                'womanliness ': 'humannness',
+                'womanhood': 'humanhood', 'Womanhood': 'Humanhood',
+                'manhood': 'humanhood',  'Manhood': 'Humanhood',
+                'maiden' : 'kiddo', 'Maiden': 'Kiddo',
+                'stableboy': 'kiddo', 'Stableboy': 'Kiddo', 'stableboys': 'kiddos', 'Stableboys': 'Kiddos',
+                ' boy ': ' kiddo ', 'Boy': 'Kiddo', ' boys': ' kiddos', 'Boys': 'Kiddos',
+                'boyfriend': 'partner', 'Boyfriend': 'Partner', 'boyfriends': 'partners', 'Boyfriends': 'Partners',
+                'daughter': 'child', 'Daughter': 'Child', 'daughters': 'children',
+                ' bro ': ' sibling ',  ' sis ': ' sibling ', ' bros ': ' siblings ',  ' sistas ': ' siblings ', 
+                'brother ': 'sibling ',  ' sister ': ' sibling ', 'brothers ': 'siblings ',  ' sisters ': ' siblings ', 'Brother ': 'Sibling ',  ' Sister ': ' Sibling ', 
+                'Brothers ': 'Siblings ',  ' Sisters ': ' Siblings ',
+                ' bro.': ' sibling.',  ' sis.': ' sibling.', ' bros.': ' siblings.',  ' sistas.': ' siblings.', 
+                'brother.': 'sibling.',  ' sister.': ' sibling.', 'brothers.': 'siblings.',  ' sisters.': ' siblings.', 'Brother.': 'Sibling.',  ' Sister.': ' Sibling.', 
+                'Brothers.': 'Siblings.',  ' Sisters.': ' Siblings.', 
+                ' bro,': ' sibling,',  ' sis,': ' sibling,', ' bros,': ' siblings,',  ' sistas,': ' siblings,',
+                'brother,': 'sibling,',  ' sister,': ' sibling,', 'brothers,': 'siblings,',  ' sisters,': ' siblings,', 'Brother,': 'Sibling,',  ' Sister,': ' Sibling,', 
+                'Brothers,': 'Siblings,',  ' Sisters,': ' Siblings,', 
+                'fraternity': 'fellowship', 'sorority': 'fellowship', 'brotherhood': 'fellowship', 'sisterhood': 'fellowship', 'girlhood': 'fellowship', 'boyhood': 'fellowship', 
+                'Fraternity': 'Fellowship', 'Sorority': 'Fellowship', 'Brotherhood': 'Fellowship', 'Sisterhood': 'Fellowship', 'Boyhood': 'Fellowship', 'Girlhood': 'Fellowship', 
+                'father': 'parent', 'Father': 'Parent', 'fathers': 'parents', 'Fathers': 'Parents',
+                ' girl ': ' kiddo ', 'Girl': 'Kiddo', ' girls': ' kiddos', 'Girls': 'Kiddos',
+                'girlfriend': 'partner', 'Girlfriend': 'Partner', 'girlfriends': 'partners', 'Girlfriends': 'Partners',
+                'guys': 'everybody', 'Guys': 'Everybody',
+                ' guy': ' folk', 'Guy ': 'Folk ', 'good-guy': 'good folk', 'bad-guy': 'bad folk', 'good-guys': 'good folks', 'bad-guys': 'bad folks',
+                'Good-guy': 'Good folk', 'Bad-guy': 'Bad folk', 'Good-guys': 'Good folks', 'Bad-guys': 'Bad folks',
+                'husband': 'spouse', 'Husband': 'Spouse',
+                'ladies and gentleman': 'everybody', 'Ladies and gentleman': 'Everybody', 'Ladies and Gentleman': 'Everybody',
+                'mankind': 'humankind', 'Mankind': 'Humankind',
+                'womankind': 'humankind', 'Womankind': 'Humankind',
+                'man-made': 'artificial', 'Man-made': 'Artificial',
+                'mother': 'parent', 'Mother': 'Parent', 
+                'mom ': 'mada ', 'dad ': 'mada ', 'mommy': 'maddy', 'daddy': 'maddy', 'mom.': 'mada.', 'dad.': 'mada.', 'mommy.': 'maddy.', 'daddy.': 'maddy.',
+                'Mom ': 'Mada ', 'Dad ': 'Mada ', 'Mommy': 'Maddy', 'Daddy': 'Maddy',
+                'grandma ' : 'grandparent', 'grandpa ': 'grandparent ',
+                'nephew': 'nibling', 'Nephew': 'Nibling',
+                'niece': 'nibling', 'Niece': 'Nibling',
+                'uncle': "parent's sibling", 'Uncle': "Parent's sibling", 
+                'Aunt': "Parent's sibling", 'her aunt': "their parent's sibling", ' aunt': " parent's sibling", 
+                'aunt,': "parent's sibling,", 'aunt.': "parent's sibling.",
+                'widow': 'surviving spouse', 'widower': 'surviving spouse', 'Widow': 'Surviving spouse', 'Widower': 'Surviving spouse', 
+                ' son ': ' child ', 'Son ': 'Child ', ' sons': ' children', ' son.': ' child.', 'Son.': 'Child.',
+                'steward ': 'flight attendant ', 'Steward ': 'Flight attendant ',
+                'stewardess ': 'flight attendant', 'Stewardess ': 'Flight attendant', 'stewardesses': 'flight attendants', 'Stewardesses': 'Flight attendants',
+                'waiter ': 'server', 'Waiter ': 'Server',
+                'waitress ': 'server', 'Waitress ': 'Server', 'waitresses': 'servers', 'Waitresses': 'Servers',
+                'wife': 'spouse', 'Wife': 'Spouse', 'wives': 'spouses', 'Wives': 'Spouses',
+                ' bishop': ' member of the clergy', 'archbishop ': 'member of the clergy ', 'clergyman': 'member of the clergy', 'clergywoman': 'member of the clergy', 
+                ' bishops': ' members of the clergy', 'archbishops': 'members of the clergy', 'clergymen': 'members of the clergy', 'clergywomen': 'members of the clergy',
+                ' Bishop': ' Member of the clergy', 'Archbishop ': 'Member of the clergy ', 'Clergyman': 'Member of the clergy', 'Clergywoman': 'Member of the clergy', 
+                ' Bishops': ' Members of the clergy', 'Archbishops': 'Members of the clergy', 'Clergymen': 'Members of the clergy', 'Clergywomen': 'Members of the clergy',
+                'deacon ': 'member of the diaconate ', 'deaconess ': 'member of the diaconate ', 'deacons': 'members of the diaconate', 'deaconesses': 'members of the diaconate',
+                'Deacon ': 'Member of the diaconate ', 'Deaconess ': 'Member of the diaconate ', 'Deacons': 'Members of the diaconate', 'Deaconesses': 'Members of the diaconate',
+                'abbot ': 'head of monastery ', 'abbots': 'heads of monastery', 'Abbot ': 'Head of monastery ', 'Abbots': 'Heads of monastery',
+                ' nun ': ' Person belonging to a religious order ', 'priest ': 'performer of sacred rites in a religious community ', ' nuns': ' person belonging to a religious order', 
+                'priests': 'members of a religious community',
+                'Nun ': 'Person belonging to a religious order ', 'Priest ': 'performers of sacred rites in a religious community ', 'Nuns': 'Person belonging to a religious order', 
+                'Priests': 'Performers of sacred rites in a religious community',
+                'Priestess': 'Performer of sacred rites in a religious community', 'Priestesses': 'Performers of sacred rites in a religious community',
+                'priestess': 'performer of sacred rites in a religious community', 'priestesses': 'performers of sacred rites in religious community',
+                ' pope ': ' head of the Roman catholic church ', 'Pope ': 'Head of the Roman catholic church ', ' Popes ': ' Heads of the Roman catholic church ', 
+                ' popes ': ' heads of the Roman catholic church ', 
+                'headmaster': 'head', 'headmistress ': 'head ', 'Headmaster': 'Head', 'Headmistress ': 'Head ',
+                'headmistresses': 'heads', 'Headmistresses': 'Heads',
+                ' her to ': ' them to ',
+                ' maid ': ' servant ',  'Maid ': 'Servant ', ' maid,': ' servant,', ' maids': ' servants', ' maid.': ' servant.', ' maid:': ' servant:', 
+                'Maids': 'Servants',
+                'butler': 'main servant', 'Butler': 'Main servant', 
+                }
+
+def neutral_converter(string, substitutions = dict_neutral): #defaults to dictionary
+    substrings = sorted(substitutions, key=len, reverse=True)
+    regex = re.compile('|'.join(map(re.escape, substrings)))
+    return regex.sub(lambda match: substitutions[match.group(0)], string)   #subn to count conversions
+
+
+
+    
+    
+#return(str(result))
